+++ src/provider/utils/set.ts
@@ -5,7 +5,7 @@ import { DBSet, DynamodbExecutor } from './dynamoDB';
 export type { DBSet };
 
 export class DynamoDBSet extends DynamodbExecutor {
-  createSet(items: string[] | number[]): DBSet {
-    return this.createSet(items);
+  createSet<T extends string[] | number[]>(items: T): DBSet<T[number], 'v2' | 'v3'> {
+    return this._createSet(items);
   }
 }